Put the parser in as a front-end to conversion
We now parse the number to ensure strict JSON compliance before passing
to the StringToDouble converter.
Also added check to ensure parsed value is not-infinite (which could
happen by supplying a number larger than the maximum representable).
Added XCode IDE project and hacked existing projects to support 64bit.
This needs to be cleaned up (along with a lot of warnings).