Update adobe::promote to use fixed width types and remove 64bit Win promote warning.
Based on discussion with sparent who said "The code should be written in terms of fixed sized types" and (regarding the warning) "I’m sure the error message is because we had cases of size_t for 32 bits and in the move to 64 bits this broke so we put in the error. Since size_t is not a distinct type this is really just going to generate an error for any 64 bit type on windows. These cases should be removed."